Don't let amazonka overwrite request timeout
It may be doing more but I found out that during the request conversion,
amazonka was setting its own timeout on the requests. This isn't
appropriate as the user may have set their own (or wanted it to inherit
from manager). This was causing long running elasticsearch calls to
timeout due to amazonka's conservative 70s timeout. Fine for most things
but snapshot/restore can take a long time.